requirements.txt
gitpython>=3.1.24
PyGithub>=1.55
python-dotenv>=0.19.2
Flask>=2.1.1
